+++ layers/Schema/packages/media/src/TypeAPIs/MediaTypeAPIInterface.php
@@ -19,6 +19,7 @@ interface MediaTypeAPIInterface
     public function getImageSrcSet(string | int $image_id, ?string $size = null): ?string;
     public function getImageSizes(string | int $image_id, ?string $size = null): ?string;
     public function getImageProperties(string | int $image_id, ?string $size = null): ?array;
-    public function getMediaElements(array $query, array $options = []): array;
-    public function getMediaElementId(object $media): string | int;
+    public function getMediaItems(array $query, array $options = []): array;
+    public function getMediaItemCount(array $query, array $options = []): int;
+    public function getMediaItemId(object $media): string | int;
 }